Large client rework
client reworked based on MVC approach

+++ src/com/metransfert/client/gui/DownloadPanel.java
@@ -1,6 +1,6 @@
 package com.metransfert.client.gui;
 
-import com.metransfert.client.transaction.TransferListener;
+import com.metransfert.client.transactionhandlers.TransferListener;
 
 import javax.swing.*;
 import java.nio.file.Path;
@@ -49,9 +49,9 @@ public class DownloadPanel extends JPanel{
         progressBar.setString("Downloading... " +percentage+" %"+ "("+throughputString+")");
     }
 
-    private long getThroughput(int currentBytes, int lastBytes){
+    private long getThroughput(long currentBytes, long lastBytes){
         long currentNano = System.nanoTime();
-        long dB = (long)currentBytes - (long)lastBytes;
+        long dB = currentBytes - lastBytes;
         long dt = currentNano - lastNano;
         long dt_s = dt/1_000_000_000L;
         if(dB != 0 && dt_s != 0)
